removed terms.firestore & term.model =>mdel
