+++ libs/model/src/lib/terms.ts
@@ -6,6 +6,7 @@ import { staticModel } from '@blockframes/utils/static-model';
 import { format } from 'date-fns';
 import { toLanguageVersionString } from '@blockframes/utils/utils';
 
+
 export function createMailTerm(terms: BucketTerm<Timestamp>[]) {
   return terms.map((term) => ({
     ...term,
@@ -52,3 +53,19 @@ export interface Term<T extends Date | firebase.firestore.Timestamp = Date> exte
 }
 
 export type TermDocument = Term<firebase.firestore.Timestamp>;
+
+
+export function createTerm(params: Partial<Term<Date>> = {}): Term<Date> {
+  return {
+    id: '',
+    contractId: '',
+    territories: [],
+    medias: [],
+    exclusive: false,
+    duration: { from: new Date(), to: new Date() },
+    licensedOriginal: null,
+    languages: {},
+    criteria: [],
+    ...params
+  }
+}
